Meachine Learning (Deploy)/requirements.txt
flask
tensorflow
numpy
Pillow
flask-cors